refactor: combine all classes to init file
